# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Meyora API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/organization": {
   "get": {
    "tags": [
     "organization"
    ],
    "summary": "Get Organization",
    "operationId": "get_organization_api_organization_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OrganizationOut"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "patch": {
    "tags": [
     "organization"
    ],
    "summary": "Update Organization",
    "operationId": "update_organization_api_organization_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OrganizationUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OrganizationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/organization/members": {
   "get": {
    "tags": [
     "organization"
    ],
    "summary": "List Members",
    "operationId": "list_members_api_organization_members_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/MemberOut"
         },
         "type": "array",
         "title": "Response List Members Api Organization Members Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/organization/members/{user_id}": {
   "patch": {
    "tags": [
     "organization"
    ],
    "summary": "Change Role",
    "operationId": "change_role_api_organization_members__user_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "User Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RoleUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MemberOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "organization"
    ],
    "summary": "Remove Member",
    "operationId": "remove_member_api_organization_members__user_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/organization/invitations": {
   "get": {
    "tags": [
     "organization"
    ],
    "summary": "List Invitations",
    "operationId": "list_invitations_api_organization_invitations_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/InvitationOut"
         },
         "type": "array",
         "title": "Response List Invitations Api Organization Invitations Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "organization"
    ],
    "summary": "Create Invitation",
    "operationId": "create_invitation_api_organization_invitations_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InvitationCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationCreated"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/organization/invitations/{invitation_id}": {
   "delete": {
    "tags": [
     "organization"
    ],
    "summary": "Revoke Invitation",
    "operationId": "revoke_invitation_api_organization_invitations__invitation_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "invitation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Invitation Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/organization/audit-logs": {
   "get": {
    "tags": [
     "organization"
    ],
    "summary": "Audit Logs",
    "operationId": "audit_logs_api_organization_audit_logs_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "entity_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 50
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Type"
      }
     },
     {
      "name": "entity_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Id"
      }
     },
     {
      "name": "actor_user_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Actor User Id"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 10000,
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 25,
       "title": "Page Size"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 200
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 50,
         "pattern": "^-?[a-z_]+$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Sort"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_AuditLogOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/pipelines": {
   "get": {
    "tags": [
     "pipelines"
    ],
    "summary": "List Pipelines",
    "operationId": "list_pipelines_api_pipelines_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/PipelineOut"
         },
         "type": "array",
         "title": "Response List Pipelines Api Pipelines Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "pipelines"
    ],
    "summary": "Create Pipeline",
    "operationId": "create_pipeline_api_pipelines_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PipelineCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PipelineOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/pipelines/{pipeline_id}": {
   "patch": {
    "tags": [
     "pipelines"
    ],
    "summary": "Update Pipeline",
    "operationId": "update_pipeline_api_pipelines__pipeline_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "pipeline_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Pipeline Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PipelineUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PipelineOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "pipelines"
    ],
    "summary": "Delete Pipeline",
    "operationId": "delete_pipeline_api_pipelines__pipeline_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "pipeline_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Pipeline Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/dashboard": {
   "get": {
    "tags": [
     "dashboard"
    ],
    "summary": "Get Dashboard",
    "operationId": "get_dashboard_api_dashboard_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "tz",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "maxLength": 64,
       "default": "Asia/Kolkata",
       "title": "Tz"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DashboardOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ActorType": {
    "type": "string",
    "enum": [
     "user",
     "ai",
     "system"
    ],
    "title": "ActorType"
   },
   "AuditLogOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "actor_user_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Actor User Id"
     },
     "actor_type": {
      "$ref": "#/components/schemas/ActorType"
     },
     "action": {
      "type": "string",
      "title": "Action"
     },
     "entity_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Entity Type"
     },
     "entity_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Entity Id"
     },
     "changes": {
      "additionalProperties": true,
      "type": "object",
      "title": "Changes"
     },
     "ip": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ip"
     },
     "request_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Request Id"
     }
    },
    "type": "object",
    "required": [
     "id",
     "created_at",
     "actor_user_id",
     "actor_type",
     "action",
     "entity_type",
     "entity_id",
     "changes",
     "ip",
     "request_id"
    ],
    "title": "AuditLogOut"
   },
...
   "DashboardOut": {
    "properties": {
     "currency": {
      "type": "string",
      "title": "Currency"
     },
     "generated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Generated At"
     },
     "leads": {
      "$ref": "#/components/schemas/LeadStats"
     },
     "pipeline": {
      "$ref": "#/components/schemas/PipelineStats"
     },
     "tasks": {
      "$ref": "#/components/schemas/TaskStats"
     },
     "my_tasks": {
      "items": {
       "$ref": "#/components/schemas/TaskOut"
      },
      "type": "array",
      "title": "My Tasks"
     },
     "upcoming_activities": {
      "items": {
       "$ref": "#/components/schemas/ActivityOut"
      },
      "type": "array",
      "title": "Upcoming Activities"
     },
     "recent_activities": {
      "items": {
       "$ref": "#/components/schemas/ActivityOut"
      },
      "type": "array",
      "title": "Recent Activities"
     },
     "recent_companies": {
      "items": {
       "$ref": "#/components/schemas/CompanyOut"
      },
      "type": "array",
      "title": "Recent Companies"
     },
     "recent_contacts": {
      "items": {
       "$ref": "#/components/schemas/ContactOut"
      },
      "type": "array",
      "title": "Recent Contacts"
     },
     "recent_deals": {
      "items": {
       "$ref": "#/components/schemas/DealOut"
      },
      "type": "array",
      "title": "Recent Deals"
     }
    },
    "type": "object",
    "required": [
     "currency",
     "generated_at",
     "leads",
     "pipeline",
     "tasks",
     "my_tasks",
     "upcoming_activities",
     "recent_activities",
     "recent_companies",
     "recent_contacts",
     "recent_deals"
    ],
    "title": "DashboardOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "InvitationCreate": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "role": {
      "$ref": "#/components/schemas/Role",
      "default": "member"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "email"
    ],
    "title": "InvitationCreate"
   },
   "InvitationCreated": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     },
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     },
     "accepted_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Accepted At"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "token": {
      "type": "string",
      "title": "Token"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "role",
     "expires_at",
     "accepted_at",
     "created_at",
     "token"
    ],
    "title": "InvitationCreated"
   },
   "InvitationOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     },
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     },
     "accepted_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Accepted At"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "role",
     "expires_at",
     "accepted_at",
     "created_at"
    ],
    "title": "InvitationOut"
   },
...
   "LeadStats": {
    "properties": {
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "new_last_30_days": {
      "type": "integer",
      "title": "New Last 30 Days"
     },
     "new_status": {
      "type": "integer",
      "title": "New Status"
     },
     "qualified": {
      "type": "integer",
      "title": "Qualified"
     },
     "converted": {
      "type": "integer",
      "title": "Converted"
     }
    },
    "type": "object",
    "required": [
     "total",
     "new_last_30_days",
     "new_status",
     "qualified",
     "converted"
    ],
    "title": "LeadStats"
   },
...
   "MemberOut": {
    "properties": {
     "user": {
      "$ref": "#/components/schemas/UserOut"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "user",
     "role",
     "created_at"
    ],
    "title": "MemberOut"
   },
...
   "OrganizationOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "default_currency": {
      "type": "string",
      "title": "Default Currency"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "slug",
     "default_currency"
    ],
    "title": "OrganizationOut"
   },
   "OrganizationUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "default_currency": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^[A-Z]{3}$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Currency"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "OrganizationUpdate"
   },
...
   "Page_AuditLogOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/AuditLogOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[AuditLogOut]"
   },
...
   "PipelineCreate": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Name"
     },
     "is_default": {
      "type": "boolean",
      "title": "Is Default",
      "default": false
     },
     "stages": {
      "items": {
       "$ref": "#/components/schemas/StageInput"
      },
      "type": "array",
      "maxItems": 20,
      "minItems": 3,
      "title": "Stages"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "name",
     "stages"
    ],
    "title": "PipelineCreate"
   },
   "PipelineOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "is_default": {
      "type": "boolean",
      "title": "Is Default"
     },
     "stages": {
      "items": {
       "$ref": "#/components/schemas/StageOut"
      },
      "type": "array",
      "title": "Stages"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "is_default",
     "stages"
    ],
    "title": "PipelineOut"
   },
   "PipelineStats": {
    "properties": {
     "open_deals": {
      "type": "integer",
      "title": "Open Deals"
     },
     "won_deals": {
      "type": "integer",
      "title": "Won Deals"
     },
     "lost_deals": {
      "type": "integer",
      "title": "Lost Deals"
     },
     "pipeline_value": {
      "type": "number",
      "title": "Pipeline Value"
     },
     "weighted_value": {
      "type": "number",
      "title": "Weighted Value"
     },
     "revenue_total": {
      "type": "number",
      "title": "Revenue Total"
     },
     "revenue_this_month": {
      "type": "number",
      "title": "Revenue This Month"
     },
     "closing_this_month": {
      "type": "integer",
      "title": "Closing This Month"
     },
     "other_currency_deals": {
      "type": "integer",
      "title": "Other Currency Deals"
     },
     "stages": {
      "items": {
       "$ref": "#/components/schemas/StageSummary"
      },
      "type": "array",
      "title": "Stages"
     }
    },
    "type": "object",
    "required": [
     "open_deals",
     "won_deals",
     "lost_deals",
     "pipeline_value",
     "weighted_value",
     "revenue_total",
     "revenue_this_month",
     "closing_this_month",
     "other_currency_deals",
     "stages"
    ],
    "title": "PipelineStats"
   },
   "PipelineUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "is_default": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Default"
     },
     "stages": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/StageInput"
        },
        "type": "array",
        "maxItems": 20,
        "minItems": 3
       },
       {
        "type": "null"
       }
      ],
      "title": "Stages"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "PipelineUpdate"
   },
...
   "Role": {
    "type": "string",
    "enum": [
     "owner",
     "admin",
     "manager",
     "member"
    ],
    "title": "Role"
   },
   "RoleUpdate": {
    "properties": {
     "role": {
      "$ref": "#/components/schemas/Role"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "role"
    ],
    "title": "RoleUpdate"
   },
   "StageInput": {
    "properties": {
     "id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Name"
     },
     "probability": {
      "type": "integer",
      "maximum": 100.0,
      "minimum": 0.0,
      "title": "Probability"
     },
     "kind": {
      "$ref": "#/components/schemas/StageKind",
      "default": "open"
     },
     "color": {
      "type": "string",
      "maxLength": 20,
      "pattern": "^[a-z]+$",
      "title": "Color",
      "default": "slate"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "name",
     "probability"
    ],
    "title": "StageInput"
   },
   "StageKind": {
    "type": "string",
    "enum": [
     "open",
     "won",
     "lost"
    ],
    "title": "StageKind"
   },
   "StageOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "position": {
      "type": "integer",
      "title": "Position"
     },
     "probability": {
      "type": "integer",
      "title": "Probability"
     },
     "kind": {
      "$ref": "#/components/schemas/StageKind"
     },
     "color": {
      "type": "string",
      "title": "Color"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "position",
     "probability",
     "kind",
     "color"
    ],
    "title": "StageOut"
   },
   "StageSummary": {
    "properties": {
     "stage_id": {
      "type": "string",
      "format": "uuid",
      "title": "Stage Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "color": {
      "type": "string",
      "title": "Color"
     },
     "count": {
      "type": "integer",
      "title": "Count"
     },
     "amount": {
      "type": "number",
      "title": "Amount"
     }
    },
    "type": "object",
    "required": [
     "stage_id",
     "name",
     "kind",
     "color",
     "count",
     "amount"
    ],
    "title": "StageSummary"
   },
...
   "TaskStats": {
    "properties": {
     "due_today": {
      "type": "integer",
      "title": "Due Today"
     },
     "overdue": {
      "type": "integer",
      "title": "Overdue"
     },
     "open": {
      "type": "integer",
      "title": "Open"
     }
    },
    "type": "object",
    "required": [
     "due_today",
     "overdue",
     "open"
    ],
    "title": "TaskStats"
   },
...
   "UserOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "full_name": {
      "type": "string",
      "title": "Full Name"
     },
     "avatar_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avatar Url"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "full_name",
     "avatar_url"
    ],
    "title": "UserOut"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}